slim demo completed
