+++ lenskit-slim/src/main/java/org/lenskit/slim/SlimModel.java
