+++ lenskit-slim/src/main/java/org/lenskit/slim/SlimModelProvider.java
